Rename globus_app.py -> app.py
